pyproject.toml
[tool.poetry]
name = "extrasss"
version = "0.1.0"
description = "hackathon 2.0"
authors = ["raushan <raushan.in@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.7"
djangorestframework = "^3.11.0"
drf-yasg = "^1.17.1"
pillow = "^7.1.2"
django-cors-headers = "^3.4.0"
hashids = "^1.2.0"
gunicorn = "^20.0.4"
python-dateutil = "^2.8.1"
DateTime = "^4.3"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
